fix: splits up translations
