+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/translation/de.ts
@@ -1,0 +1,117 @@
+import type { TranslationKey } from './en'
+
+export const de: Record<TranslationKey, string> = {
+    align_bottom: 'Unten',
+    align_center: 'Zentriert',
+    align_horizontal: 'Horizontale Ausrichtung',
+    align_justify: 'Blocksatz',
+    align_left: 'Linksbündig',
+    align_middle: 'Mitte',
+    align_right: 'Rechtsbündig',
+    align_top: 'Oben',
+    align_vertical: 'Vertikale Ausrichtung',
+    alignment: 'Ausrichtung',
+    block_quote_insert: 'Blockzitat',
+    border: 'Rahmen',
+    button_apply: 'Anwenden',
+    button_cancel: 'Abbrechen',
+    button_insert: 'Einfügen',
+    button_update: 'Aktualisieren',
+    caption: 'Beschriftung',
+    class: 'Klasse',
+    classes: 'Klassen',
+    choice_no: 'Nein',
+    choice_yes: 'Ja',
+    div_create: 'Div erstellen',
+    div_edit: 'Div bearbeiten',
+    div_field_lang: 'Sprachcode',
+    div_field_title: 'Beratender Titel',
+    div_remove: 'Div entfernen',
+    div_update: 'Div-Container bearbeiten',
+    format_paragraph: 'Absatzformat',
+    format_remove: 'Formatierung entfernen',
+    height: 'Höhe',
+    history_redo: 'Wiederherstellen',
+    history_undo: 'Rückgängig machen',
+    horizontal_line_insert: 'Horizontale Linie einfügen',
+    id: 'ID',
+    indent_decrease: 'Einzug verkleinen',
+    indent_increase: 'Einzug vergrößern',
+    link: 'Link',
+    link_anchor: 'Anker',
+    link_anchor_edit: 'Anker bearbeiten',
+    link_anchor_name: 'Ankername',
+    link_anchor_properties: 'Ankereigenschaften',
+    link_anchor_remove: 'Anker entfernen',
+    link_anchor_select: 'Anker auswählen',
+    link_edit: 'Link bearbeiten',
+    link_email_address: 'E-Mail-Adresse',
+    link_email_body: 'Nachrichtentext',
+    link_email_subject: 'Betreff',
+    link_no_anchors: 'Keine Anker im Dokument verfügbar',
+    link_phone_number: 'Telefonnummer',
+    link_target: 'Zielseite',
+    link_target_new_window: 'Neues Fenster (_blank)',
+    link_target_same_window: 'Gleiches Fenster (_self)',
+    link_type: 'Link-Typ',
+    link_type_anchor: 'Link zu Anker im Text',
+    link_type_email: 'E-Mail',
+    link_type_phone: 'Telefon',
+    link_type_url: 'URL',
+    link_unlink: 'Link entfernen',
+    link_url: 'URL',
+    list_bulleted_create: 'Aufzählungsliste einfügen/entfernen',
+    list_numbered_create: 'Nummerierte Liste einfügen/entfernen',
+    modal_close: 'Schließen',
+    mode_source_code: 'Quellcode anzeigen',
+    mode_source_code_hide: 'Quellcode ausblenden',
+    padding: 'Innenabstand',
+    select: 'Auswählen',
+    spacing: 'Abstand',
+    special_characters: 'Sonderzeichen',
+    style: 'Stijl',
+    style_inline: 'Inline-Stil',
+    styles_format: 'Formatvorlagen',
+    summary: 'Zusammenfassung',
+    table_cell: 'Zelle',
+    table_cell_clear: 'Zelle(n) leeren',
+    table_cell_insert_after: 'Zelle danach einfügen',
+    table_cell_insert_before: 'Zelle davor einfügen',
+    table_cell_merge: 'Zellen verbinden',
+    table_cell_properties: 'Zelleneigenschaften',
+    table_cell_span_columns: 'Spalten zusammenfassen',
+    table_cell_span_rows: 'Zeilen zusammenfassen',
+    table_cell_split: 'Zelle teilen',
+    table_cell_type: 'Zellentyp',
+    table_cell_type_data: 'Daten',
+    table_cell_type_header: 'Überschrift',
+    table_column: 'Spalte',
+    table_column_delete: 'Spalte(n) löschen',
+    table_column_insert_after: 'Spalte danach einfügen',
+    table_column_insert_before: 'Spalte davor einfügen',
+    table_columns: 'Spalten',
+    table_delete: 'Tabelle löschen',
+    table_headers: 'Überschriften',
+    table_headers_both: 'Beide',
+    table_headers_first_column: 'Erste Spalte',
+    table_headers_first_row: 'Erste Zeile',
+    table_insert: 'Tabelle einfügen',
+    table_properties: 'Tabelleneigenschaften',
+    table_row: 'Zeile',
+    table_row_delete: 'Zeile(n) löschen',
+    table_row_insert_after: 'Zeile danach einfügen',
+    table_row_insert_before: 'Zeile davor einfügen',
+    table_rows: 'Zeile',
+    text_bold: 'Fett',
+    text_italic: 'Kursiv',
+    text_strike: 'Durchgestrichen',
+    text_subscript: 'Tiefgestellt',
+    text_superscript: 'Hochgestellt',
+    tools_maximize: 'Maximieren',
+    tools_minimize: 'Minimieren',
+    tools_show_Blocks: 'Blöcke anzeigen',
+    width: 'Breite',
+    word_wrap: 'Zeilenumbruch'
+} as const
+
+export default de
