+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/translation/en.ts
@@ -1,0 +1,116 @@
+const en = {
+    align_bottom: 'Bottom',
+    align_center: 'Center',
+    align_horizontal: 'Horizontal alignment',
+    align_justify: 'Justify',
+    align_left: 'Align left',
+    align_middle: 'Middle',
+    align_right: 'Align Right',
+    align_top: 'Top',
+    align_vertical: 'Vertical alignment',
+    alignment: 'Alignment',
+    block_quote_insert: 'Block Quote',
+    border: 'Border',
+    button_apply: 'Apply',
+    button_cancel: 'Cancel',
+    button_insert: 'Insert',
+    button_update: 'Update',
+    caption: 'Caption',
+    class: 'Class',
+    classes: 'Classes',
+    choice_no: 'No',
+    choice_yes: 'Yes',
+    div_create: 'Create Div',
+    div_edit: 'Edit Div',
+    div_field_lang: 'Language code',
+    div_field_title: 'Advisory Title',
+    div_remove: 'Remove Div',
+    div_update: 'Edit Div Container',
+    format_paragraph: 'Paragraph format',
+    format_remove: 'Remove Format',
+    height: 'Height',
+    history_redo: 'Redo',
+    history_undo: 'Undo',
+    horizontal_line_insert: 'Insert Horizontal Line',
+    id: 'ID',
+    indent_decrease: 'Decrease Indent',
+    indent_increase: 'Increase Indent',
+    link: 'Link',
+    link_anchor: 'Anchor',
+    link_anchor_edit: 'Edit Anchor',
+    link_anchor_name: 'Anchor Name',
+    link_anchor_properties: 'Anchor Properties',
+    link_anchor_remove: 'Remove Anchor',
+    link_anchor_select: 'Select Anchor',
+    link_edit: 'Edit Link',
+    link_email_address: 'E-mail Address',
+    link_email_body: 'Message Body',
+    link_email_subject: 'Message Subject',
+    link_no_anchors: 'No anchors available in the document',
+    link_phone_number: 'Phone Number',
+    link_target: 'Target',
+    link_target_new_window: 'New Window (_blank)',
+    link_target_same_window: 'Same Window (_self)',
+    link_type: 'Link Type',
+    link_type_anchor: 'Link to anchor in the text',
+    link_type_email: 'E-mail',
+    link_type_phone: 'Phone',
+    link_type_url: 'URL',
+    link_unlink: 'Unlink',
+    link_url: 'URL',
+    list_bulleted_create: 'Insert/Remove Bulleted List',
+    list_numbered_create: 'Insert/Remove Numbered List',
+    modal_close: 'Close',
+    mode_source_code: 'Show Source',
+    mode_source_code_hide: 'Hide Source',
+    padding: 'Padding',
+    select: 'Select',
+    spacing: 'Spacing',
+    special_characters: 'Special Characters',
+    style: 'Style',
+    style_inline: 'Inline Style',
+    styles_format: 'Formatting styles',
+    summary: 'Summary',
+    table_cell: 'Cell',
+    table_cell_clear: 'Clear cell(s)',
+    table_cell_insert_after: 'Insert cell after',
+    table_cell_insert_before: 'Insert cell before',
+    table_cell_merge: 'Merge cells',
+    table_cell_properties: 'Cell properties',
+    table_cell_span_columns: 'Columns span',
+    table_cell_span_rows: 'Rows span',
+    table_cell_split: 'Split cell',
+    table_cell_type: 'Cell type',
+    table_cell_type_data: 'Data',
+    table_cell_type_header: 'Header',
+    table_column: 'Column',
+    table_column_delete: 'Delete column(s)',
+    table_column_insert_after: 'Insert column after',
+    table_column_insert_before: 'Insert column before',
+    table_columns: 'Columns',
+    table_delete: 'Delete table',
+    table_headers: 'Headers',
+    table_headers_both: 'Both',
+    table_headers_first_column: 'First column',
+    table_headers_first_row: 'First row',
+    table_insert: 'Insert table',
+    table_properties: 'Table properties',
+    table_row: 'Row',
+    table_row_delete: 'Delete row(s)',
+    table_row_insert_after: 'Insert row after',
+    table_row_insert_before: 'Insert row before',
+    table_rows: 'Row',
+    text_bold: 'Bold',
+    text_italic: 'Italic',
+    text_strike: 'Strike Through',
+    text_subscript: 'Subscript',
+    text_superscript: 'Superscript',
+    tools_maximize: 'Maximize',
+    tools_minimize: 'Minimize',
+    tools_show_Blocks: 'Show Blocks',
+    width: 'Width',
+    word_wrap: 'Word wrap'
+} as const
+
+export type TranslationKey = keyof typeof en
+export default en
