+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/translation/fr.ts
@@ -1,0 +1,117 @@
+import type { TranslationKey } from './en'
+
+export const fr: Record<TranslationKey, string> = {
+    align_bottom: 'Bas',
+    align_center: 'Centrer',
+    align_horizontal: 'Alignement horizontal',
+    align_justify: 'Justifier',
+    align_left: 'Aligner à gauche',
+    align_middle: 'Milieu',
+    align_right: 'Aligner à droite',
+    align_top: 'Haut',
+    align_vertical: 'Alignement vertical',
+    alignment: 'Alignement',
+    block_quote_insert: 'Bloc de citation',
+    border: 'Bordure',
+    button_apply: 'Appliquer',
+    button_cancel: 'Annuler',
+    button_insert: 'Insérer',
+    button_update: 'Mettre à jour',
+    caption: 'Légende',
+    class: 'Classe',
+    classes: 'Classes',
+    choice_no: 'Non',
+    choice_yes: 'Oui',
+    div_create: 'Créer un Div',
+    div_edit: 'Modifier la Div',
+    div_field_lang: 'Code de langue',
+    div_field_title: 'Titre informatif',
+    div_remove: 'Supprimer la Div',
+    div_update: 'Modifier le conteneur Div',
+    format_paragraph: 'Format de paragraphe',
+    format_remove: 'Supprimer la mise en forme',
+    height: 'Hauteur',
+    history_redo: 'Rétablir',
+    history_undo: 'Annuler',
+    horizontal_line_insert: 'Insérer une ligne horizontale',
+    id: 'ID',
+    indent_decrease: 'Diminuer le retrait',
+    indent_increase: 'Augmenter le retrait',
+    link: 'Lien',
+    link_anchor: 'Ancre',
+    link_anchor_edit: "Modifier l'ancre",
+    link_anchor_name: "Nom de l'ancre",
+    link_anchor_properties: "Propriétés de l'ancre",
+    link_anchor_remove: "Supprimer l'ancre",
+    link_anchor_select: 'Sélectionner une ancre',
+    link_edit: 'Modifier le lien',
+    link_email_address: 'Adresse e-mail',
+    link_email_body: 'Corps du message',
+    link_email_subject: 'Objet du message',
+    link_no_anchors: 'Aucune ancre disponible dans le document',
+    link_phone_number: 'Numéro de téléphone',
+    link_target: 'Cible',
+    link_target_new_window: 'Nouvelle fenêtre (_blank)',
+    link_target_same_window: 'Même fenêtre (_self)',
+    link_type: 'Type de lien',
+    link_type_anchor: "Lien vers l'ancre dans le texte",
+    link_type_email: 'E-mail',
+    link_type_phone: 'Téléphone',
+    link_type_url: 'URL',
+    link_unlink: 'Supprimer le lien',
+    link_url: 'URL',
+    list_bulleted_create: 'Insérer/Supprimer une liste à puces',
+    list_numbered_create: 'Insérer/Supprimer une liste numérotée',
+    modal_close: 'Fermer',
+    mode_source_code: 'Afficher la source',
+    mode_source_code_hide: 'Masquer la source',
+    padding: 'Marge intérieure',
+    select: 'Sélectionner',
+    spacing: 'Espacement',
+    special_characters: 'Caractères spéciaux',
+    style: 'Style',
+    style_inline: 'Style en ligne',
+    styles_format: 'Styles de mise en forme',
+    summary: 'Résumé',
+    table_cell: 'Cellule',
+    table_cell_clear: 'Effacer la/les cellule(s)',
+    table_cell_insert_after: 'Insérer une cellule après',
+    table_cell_insert_before: 'Insérer une cellule avant',
+    table_cell_merge: 'Fusionner les cellules',
+    table_cell_properties: 'Propriétés de la cellule',
+    table_cell_span_columns: 'Fusion de colonnes',
+    table_cell_span_rows: 'Fusion de lignes',
+    table_cell_split: 'Fractionner la cellule',
+    table_cell_type: 'Type de cellule',
+    table_cell_type_data: 'Données',
+    table_cell_type_header: 'En-tête',
+    table_column: 'Colonne',
+    table_column_delete: 'Supprimer la/les colonne(s)',
+    table_column_insert_after: 'Insérer une colonne après',
+    table_column_insert_before: 'Insérer une colonne avant',
+    table_columns: 'Colonnes',
+    table_delete: 'Supprimer le tableau',
+    table_headers: 'En-têtes',
+    table_headers_both: 'Les deux',
+    table_headers_first_column: 'Première colonne',
+    table_headers_first_row: 'Première ligne',
+    table_insert: 'Insérer un tableau',
+    table_properties: 'Propriétés du tableau',
+    table_row: 'Ligne',
+    table_row_delete: 'Supprimer la/les ligne(s)',
+    table_row_insert_after: 'Insérer une ligne après',
+    table_row_insert_before: 'Insérer une ligne avant',
+    table_rows: 'Ligne',
+    text_bold: 'Gras',
+    text_italic: 'Italique',
+    text_strike: 'Barré',
+    text_subscript: 'Indice',
+    text_superscript: 'Exposant',
+    tools_maximize: 'Agrandir',
+    tools_minimize: 'Réduire',
+    tools_show_Blocks: 'Afficher les blocs',
+    width: 'Largeur',
+    word_wrap: 'Retour à la ligne'
+} as const
+
+export default fr
