+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/translation/nl.ts
@@ -1,0 +1,117 @@
+import type { TranslationKey } from './en'
+
+export const nl: Record<TranslationKey, string> = {
+    align_bottom: 'Onderaan',
+    align_center: 'Centreren',
+    align_horizontal: 'Horizontale uitlijning',
+    align_justify: 'Uitvullen',
+    align_left: 'Links uitlijnen',
+    align_middle: 'Midden',
+    align_right: 'Rechts uitlijnen',
+    align_top: 'Bovenaan',
+    align_vertical: 'Verticale uitlijning',
+    alignment: 'Uitlijning',
+    block_quote_insert: 'Citaatblok',
+    border: 'Rand',
+    button_apply: 'Toepassen',
+    button_cancel: 'Annuleren',
+    button_insert: 'Invoegen',
+    button_update: 'Bijwerken',
+    caption: 'Bijschrift',
+    class: 'Klasse',
+    classes: 'Classes',
+    choice_no: 'Nee',
+    choice_yes: 'Ja',
+    div_create: 'Div maken',
+    div_edit: 'Div bewerken',
+    div_field_lang: 'Taalcode',
+    div_field_title: 'Adviserende titel',
+    div_remove: 'Div verwijderen',
+    div_update: 'Div-container aanpassen',
+    format_paragraph: 'Paragraafopmaak',
+    format_remove: 'Opmaak verwijderen',
+    height: 'Hoogte',
+    history_redo: 'Opnieuw uitvoeren',
+    history_undo: 'Ongedaan maken',
+    horizontal_line_insert: 'Horizontale lijn invoegen',
+    id: 'ID',
+    indent_decrease: 'Inspringing verkleinen',
+    indent_increase: 'Inspringing vergroten',
+    link: 'Link',
+    link_anchor: 'Anker',
+    link_anchor_edit: 'Anker bewerken',
+    link_anchor_name: 'Ankernaam',
+    link_anchor_properties: 'Ankereigenschappen',
+    link_anchor_remove: 'Anker verwijderen',
+    link_anchor_select: 'Selecteer anker',
+    link_edit: 'Link bewerken',
+    link_email_address: 'E-mailadres',
+    link_email_body: 'Berichtinhoud',
+    link_email_subject: 'Onderwerp',
+    link_no_anchors: 'Geen ankers beschikbaar in het document',
+    link_phone_number: 'Telefoonnummer',
+    link_target: 'Doel',
+    link_target_new_window: 'Nieuw venster (_blank)',
+    link_target_same_window: 'Zelfde venster (_self)',
+    link_type: 'Link type',
+    link_type_anchor: 'Link naar anker in de tekst',
+    link_type_email: 'E-mail',
+    link_type_phone: 'Telefoon',
+    link_type_url: 'URL',
+    link_unlink: 'Link verwijderen',
+    link_url: 'URL',
+    list_bulleted_create: 'Lijst met opsommingstekens invoegen/verwijderen',
+    list_numbered_create: 'Genummerde lijst invoegen/verwijderen',
+    modal_close: 'Sluiten',
+    mode_source_code: 'Broncode tonen',
+    mode_source_code_hide: 'Broncode verbergen',
+    padding: 'Opvulling',
+    select: 'Selecteren',
+    spacing: 'Afstand',
+    special_characters: 'Speciale tekens',
+    style: 'Stijl',
+    style_inline: 'Inline stijl',
+    styles_format: 'Opmaakstijlen',
+    summary: 'Samenvatting',
+    table_cell: 'Cel',
+    table_cell_clear: 'Cel(len) leegmaken',
+    table_cell_insert_after: 'Cel invoegen na',
+    table_cell_insert_before: 'Cel invoegen voor',
+    table_cell_merge: 'Cellen samenvoegen',
+    table_cell_properties: 'Celeigenschappen',
+    table_cell_span_columns: 'Kolommen samenvoegen',
+    table_cell_span_rows: 'Rijen samenvoegen',
+    table_cell_split: 'Cel splitsen',
+    table_cell_type: 'Celtype',
+    table_cell_type_data: 'Data',
+    table_cell_type_header: 'Koptekst',
+    table_column: 'Kolom',
+    table_column_delete: 'Kolom(men) verwijderen',
+    table_column_insert_after: 'Kolom invoegen na',
+    table_column_insert_before: 'Kolom invoegen voor',
+    table_columns: 'Kolommen',
+    table_delete: 'Tabel verwijderen',
+    table_headers: 'Kopteksten',
+    table_headers_both: 'Beide',
+    table_headers_first_column: 'Eerste kolom',
+    table_headers_first_row: 'Eerste rij',
+    table_insert: 'Tabel invoegen',
+    table_properties: 'Tabeleigenschappen',
+    table_row: 'Rij',
+    table_row_delete: 'Rij(en) verwijderen',
+    table_row_insert_after: 'Rij invoegen na',
+    table_row_insert_before: 'Rij invoegen voor',
+    table_rows: 'Rij',
+    text_bold: 'Vet',
+    text_italic: 'Cursief',
+    text_strike: 'Doorhalen',
+    text_subscript: 'Subscript',
+    text_superscript: 'Superscript',
+    tools_maximize: 'Maximaliseren',
+    tools_minimize: 'Minimaliseren',
+    tools_show_Blocks: 'Blokken tonen',
+    width: 'Breedte',
+    word_wrap: 'Regelterugloop'
+} as const
+
+export default nl
