+++ EMS/admin-ui-bundle/assets/src/core/components/tiptap/translations.ts
@@ -1,466 +1,20 @@
-export type Locale = 'en' | 'nl' | 'fr' | 'de'
-const LOCALES: readonly Locale[] = ['en', 'nl', 'fr', 'de']
+import type { TranslationKey } from './translation/en'
 
-export const isTransLocale = (v: string): v is Locale => (LOCALES as readonly string[]).includes(v)
+import en from './translation/en'
+import nl from './translation/nl'
+import fr from './translation/fr'
+import de from './translation/de'
+
+export type { TranslationKey }
+
+const translations = { en, nl, fr, de } as const satisfies Record<
+    string,
+    Record<TranslationKey, string>
+>
+
+export type Locale = keyof typeof translations
+
+export const isTransLocale = (v: string): v is Locale => v in translations
 
 export const trans = (locale: Locale, key: TranslationKey): string =>
-    translations[locale]?.[key] ?? translations.en[key] ?? key
-
-const en = {
-    align_bottom: 'Bottom',
-    align_center: 'Center',
-    align_horizontal: 'Horizontal alignment',
-    align_justify: 'Justify',
-    align_left: 'Align left',
-    align_middle: 'Middle',
-    align_right: 'Align Right',
-    align_top: 'Top',
-    align_vertical: 'Vertical alignment',
-    alignment: 'Alignment',
-    block_quote_insert: 'Block Quote',
-    border: 'Border',
-    button_apply: 'Apply',
-    button_cancel: 'Cancel',
-    button_insert: 'Insert',
-    button_update: 'Update',
-    caption: 'Caption',
-    class: 'Class',
-    classes: 'Classes',
-    choice_no: 'No',
-    choice_yes: 'Yes',
-    div_create: 'Create Div',
-    div_edit: 'Edit Div',
-    div_field_lang: 'Language code',
-    div_field_title: 'Advisory Title',
-    div_remove: 'Remove Div',
-    div_update: 'Edit Div Container',
-    format_paragraph: 'Paragraph format',
-    format_remove: 'Remove Format',
-    height: 'Height',
-    history_redo: 'Redo',
-    history_undo: 'Undo',
-    horizontal_line_insert: 'Insert Horizontal Line',
-    id: 'ID',
-    indent_decrease: 'Decrease Indent',
-    indent_increase: 'Increase Indent',
-    link: 'Link',
-    link_anchor: 'Anchor',
-    link_anchor_edit: 'Edit Anchor',
-    link_anchor_name: 'Anchor Name',
-    link_anchor_properties: 'Anchor Properties',
-    link_anchor_remove: 'Remove Anchor',
-    link_anchor_select: 'Select Anchor',
-    link_edit: 'Edit Link',
-    link_email_address: 'E-mail Address',
-    link_email_body: 'Message Body',
-    link_email_subject: 'Message Subject',
-    link_no_anchors: 'No anchors available in the document',
-    link_phone_number: 'Phone Number',
-    link_target: 'Target',
-    link_target_new_window: 'New Window (_blank)',
-    link_target_same_window: 'Same Window (_self)',
-    link_type: 'Link Type',
-    link_type_anchor: 'Link to anchor in the text',
-    link_type_email: 'E-mail',
-    link_type_phone: 'Phone',
-    link_type_url: 'URL',
-    link_unlink: 'Unlink',
-    link_url: 'URL',
-    list_bulleted_create: 'Insert/Remove Bulleted List',
-    list_numbered_create: 'Insert/Remove Numbered List',
-    modal_close: 'Close',
-    mode_source_code: 'Show Source',
-    mode_source_code_hide: 'Hide Source',
-    padding: 'Padding',
-    select: 'Select',
-    spacing: 'Spacing',
-    special_characters: 'Special Characters',
-    style: 'Style',
-    style_inline: 'Inline Style',
-    styles_format: 'Formatting styles',
-    summary: 'Summary',
-    table_cell: 'Cell',
-    table_cell_clear: 'Clear cell(s)',
-    table_cell_insert_after: 'Insert cell after',
-    table_cell_insert_before: 'Insert cell before',
-    table_cell_merge: 'Merge cells',
-    table_cell_properties: 'Cell properties',
-    table_cell_span_columns: 'Columns span',
-    table_cell_span_rows: 'Rows span',
-    table_cell_split: 'Split cell',
-    table_cell_type: 'Cell type',
-    table_cell_type_data: 'Data',
-    table_cell_type_header: 'Header',
-    table_column: 'Column',
-    table_column_delete: 'Delete column(s)',
-    table_column_insert_after: 'Insert column after',
-    table_column_insert_before: 'Insert column before',
-    table_columns: 'Columns',
-    table_delete: 'Delete table',
-    table_headers: 'Headers',
-    table_headers_both: 'Both',
-    table_headers_first_column: 'First column',
-    table_headers_first_row: 'First row',
-    table_insert: 'Insert table',
-    table_properties: 'Table properties',
-    table_row: 'Row',
-    table_row_delete: 'Delete row(s)',
-    table_row_insert_after: 'Insert row after',
-    table_row_insert_before: 'Insert row before',
-    table_rows: 'Row',
-    text_bold: 'Bold',
-    text_italic: 'Italic',
-    text_strike: 'Strike Through',
-    text_subscript: 'Subscript',
-    text_superscript: 'Superscript',
-    tools_maximize: 'Maximize',
-    tools_minimize: 'Minimize',
-    tools_show_Blocks: 'Show Blocks',
-    width: 'Width',
-    word_wrap: 'Word wrap'
-} as const
-
-export type TranslationKey = keyof typeof en
-
-const translations: Record<Locale, Record<TranslationKey, string>> = {
-    en,
-    nl: {
-        align_bottom: 'Onderaan',
-        align_center: 'Centreren',
-        align_horizontal: 'Horizontale uitlijning',
-        align_justify: 'Uitvullen',
-        align_left: 'Links uitlijnen',
-        align_middle: 'Midden',
-        align_right: 'Rechts uitlijnen',
-        align_top: 'Bovenaan',
-        align_vertical: 'Verticale uitlijning',
-        alignment: 'Uitlijning',
-        block_quote_insert: 'Citaatblok',
-        border: 'Rand',
-        button_apply: 'Toepassen',
-        button_cancel: 'Annuleren',
-        button_insert: 'Invoegen',
-        button_update: 'Bijwerken',
-        caption: 'Bijschrift',
-        class: 'Klasse',
-        classes: 'Classes',
-        choice_no: 'Nee',
-        choice_yes: 'Ja',
-        div_create: 'Div maken',
-        div_edit: 'Div bewerken',
-        div_field_lang: 'Taalcode',
-        div_field_title: 'Adviserende titel',
-        div_remove: 'Div verwijderen',
-        div_update: 'Div-container aanpassen',
-        format_paragraph: 'Paragraafopmaak',
-        format_remove: 'Opmaak verwijderen',
-        height: 'Hoogte',
-        history_redo: 'Opnieuw uitvoeren',
-        history_undo: 'Ongedaan maken',
-        horizontal_line_insert: 'Horizontale lijn invoegen',
-        id: 'ID',
-        indent_decrease: 'Inspringing verkleinen',
-        indent_increase: 'Inspringing vergroten',
-        link: 'Link',
-        link_anchor: 'Anker',
-        link_anchor_edit: 'Anker bewerken',
-        link_anchor_name: 'Ankernaam',
-        link_anchor_properties: 'Ankereigenschappen',
-        link_anchor_remove: 'Anker verwijderen',
-        link_anchor_select: 'Selecteer anker',
-        link_edit: 'Link bewerken',
-        link_email_address: 'E-mailadres',
-        link_email_body: 'Berichtinhoud',
-        link_email_subject: 'Onderwerp',
-        link_no_anchors: 'Geen ankers beschikbaar in het document',
-        link_phone_number: 'Telefoonnummer',
-        link_target: 'Doel',
-        link_target_new_window: 'Nieuw venster (_blank)',
-        link_target_same_window: 'Zelfde venster (_self)',
-        link_type: 'Link type',
-        link_type_anchor: 'Link naar anker in de tekst',
-        link_type_email: 'E-mail',
-        link_type_phone: 'Telefoon',
-        link_type_url: 'URL',
-        link_unlink: 'Link verwijderen',
-        link_url: 'URL',
-        list_bulleted_create: 'Lijst met opsommingstekens invoegen/verwijderen',
-        list_numbered_create: 'Genummerde lijst invoegen/verwijderen',
-        modal_close: 'Sluiten',
-        mode_source_code: 'Broncode tonen',
-        mode_source_code_hide: 'Broncode verbergen',
-        padding: 'Opvulling',
-        select: 'Selecteren',
-        spacing: 'Afstand',
-        special_characters: 'Speciale tekens',
-        style: 'Stijl',
-        style_inline: 'Inline stijl',
-        styles_format: 'Opmaakstijlen',
-        summary: 'Samenvatting',
-        table_cell: 'Cel',
-        table_cell_clear: 'Cel(len) leegmaken',
-        table_cell_insert_after: 'Cel invoegen na',
-        table_cell_insert_before: 'Cel invoegen voor',
-        table_cell_merge: 'Cellen samenvoegen',
-        table_cell_properties: 'Celeigenschappen',
-        table_cell_span_columns: 'Kolommen samenvoegen',
-        table_cell_span_rows: 'Rijen samenvoegen',
-        table_cell_split: 'Cel splitsen',
-        table_cell_type: 'Celtype',
-        table_cell_type_data: 'Data',
-        table_cell_type_header: 'Koptekst',
-        table_column: 'Kolom',
-        table_column_delete: 'Kolom(men) verwijderen',
-        table_column_insert_after: 'Kolom invoegen na',
-        table_column_insert_before: 'Kolom invoegen voor',
-        table_columns: 'Kolommen',
-        table_delete: 'Tabel verwijderen',
-        table_headers: 'Kopteksten',
-        table_headers_both: 'Beide',
-        table_headers_first_column: 'Eerste kolom',
-        table_headers_first_row: 'Eerste rij',
-        table_insert: 'Tabel invoegen',
-        table_properties: 'Tabeleigenschappen',
-        table_row: 'Rij',
-        table_row_delete: 'Rij(en) verwijderen',
-        table_row_insert_after: 'Rij invoegen na',
-        table_row_insert_before: 'Rij invoegen voor',
-        table_rows: 'Rij',
-        text_bold: 'Vet',
-        text_italic: 'Cursief',
-        text_strike: 'Doorhalen',
-        text_subscript: 'Subscript',
-        text_superscript: 'Superscript',
-        tools_maximize: 'Maximaliseren',
-        tools_minimize: 'Minimaliseren',
-        tools_show_Blocks: 'Blokken tonen',
-        width: 'Breedte',
-        word_wrap: 'Regelterugloop'
-    },
-    fr: {
-        align_bottom: 'Bas',
-        align_center: 'Centrer',
-        align_horizontal: 'Alignement horizontal',
-        align_justify: 'Justifier',
-        align_left: 'Aligner à gauche',
-        align_middle: 'Milieu',
-        align_right: 'Aligner à droite',
-        align_top: 'Haut',
-        align_vertical: 'Alignement vertical',
-        alignment: 'Alignement',
-        block_quote_insert: 'Bloc de citation',
-        border: 'Bordure',
-        button_apply: 'Appliquer',
-        button_cancel: 'Annuler',
-        button_insert: 'Insérer',
-        button_update: 'Mettre à jour',
-        caption: 'Légende',
-        class: 'Classe',
-        classes: 'Classes',
-        choice_no: 'Non',
-        choice_yes: 'Oui',
-        div_create: 'Créer un Div',
-        div_edit: 'Modifier la Div',
-        div_field_lang: 'Code de langue',
-        div_field_title: 'Titre informatif',
-        div_remove: 'Supprimer la Div',
-        div_update: 'Modifier le conteneur Div',
-        format_paragraph: 'Format de paragraphe',
-        format_remove: 'Supprimer la mise en forme',
-        height: 'Hauteur',
-        history_redo: 'Rétablir',
-        history_undo: 'Annuler',
-        horizontal_line_insert: 'Insérer une ligne horizontale',
-        id: 'ID',
-        indent_decrease: 'Diminuer le retrait',
-        indent_increase: 'Augmenter le retrait',
-        link: 'Lien',
-        link_anchor: 'Ancre',
-        link_anchor_edit: "Modifier l'ancre",
-        link_anchor_name: "Nom de l'ancre",
-        link_anchor_properties: "Propriétés de l'ancre",
-        link_anchor_remove: "Supprimer l'ancre",
-        link_anchor_select: 'Sélectionner une ancre',
-        link_edit: 'Modifier le lien',
-        link_email_address: 'Adresse e-mail',
-        link_email_body: 'Corps du message',
-        link_email_subject: 'Objet du message',
-        link_no_anchors: 'Aucune ancre disponible dans le document',
-        link_phone_number: 'Numéro de téléphone',
-        link_target: 'Cible',
-        link_target_new_window: 'Nouvelle fenêtre (_blank)',
-        link_target_same_window: 'Même fenêtre (_self)',
-        link_type: 'Type de lien',
-        link_type_anchor: "Lien vers l'ancre dans le texte",
-        link_type_email: 'E-mail',
-        link_type_phone: 'Téléphone',
-        link_type_url: 'URL',
-        link_unlink: 'Supprimer le lien',
-        link_url: 'URL',
-        list_bulleted_create: 'Insérer/Supprimer une liste à puces',
-        list_numbered_create: 'Insérer/Supprimer une liste numérotée',
-        modal_close: 'Fermer',
-        mode_source_code: 'Afficher la source',
-        mode_source_code_hide: 'Masquer la source',
-        padding: 'Marge intérieure',
-        select: 'Sélectionner',
-        spacing: 'Espacement',
-        special_characters: 'Caractères spéciaux',
-        style: 'Style',
-        style_inline: 'Style en ligne',
-        styles_format: 'Styles de mise en forme',
-        summary: 'Résumé',
-        table_cell: 'Cellule',
-        table_cell_clear: 'Effacer la/les cellule(s)',
-        table_cell_insert_after: 'Insérer une cellule après',
-        table_cell_insert_before: 'Insérer une cellule avant',
-        table_cell_merge: 'Fusionner les cellules',
-        table_cell_properties: 'Propriétés de la cellule',
-        table_cell_span_columns: 'Fusion de colonnes',
-        table_cell_span_rows: 'Fusion de lignes',
-        table_cell_split: 'Fractionner la cellule',
-        table_cell_type: 'Type de cellule',
-        table_cell_type_data: 'Données',
-        table_cell_type_header: 'En-tête',
-        table_column: 'Colonne',
-        table_column_delete: 'Supprimer la/les colonne(s)',
-        table_column_insert_after: 'Insérer une colonne après',
-        table_column_insert_before: 'Insérer une colonne avant',
-        table_columns: 'Colonnes',
-        table_delete: 'Supprimer le tableau',
-        table_headers: 'En-têtes',
-        table_headers_both: 'Les deux',
-        table_headers_first_column: 'Première colonne',
-        table_headers_first_row: 'Première ligne',
-        table_insert: 'Insérer un tableau',
-        table_properties: 'Propriétés du tableau',
-        table_row: 'Ligne',
-        table_row_delete: 'Supprimer la/les ligne(s)',
-        table_row_insert_after: 'Insérer une ligne après',
-        table_row_insert_before: 'Insérer une ligne avant',
-        table_rows: 'Ligne',
-        text_bold: 'Gras',
-        text_italic: 'Italique',
-        text_strike: 'Barré',
-        text_subscript: 'Indice',
-        text_superscript: 'Exposant',
-        tools_maximize: 'Agrandir',
-        tools_minimize: 'Réduire',
-        tools_show_Blocks: 'Afficher les blocs',
-        width: 'Largeur',
-        word_wrap: 'Retour à la ligne'
-    },
-    de: {
-        align_bottom: 'Unten',
-        align_center: 'Zentriert',
-        align_horizontal: 'Horizontale Ausrichtung',
-        align_justify: 'Blocksatz',
-        align_left: 'Linksbündig',
-        align_middle: 'Mitte',
-        align_right: 'Rechtsbündig',
-        align_top: 'Oben',
-        align_vertical: 'Vertikale Ausrichtung',
-        alignment: 'Ausrichtung',
-        block_quote_insert: 'Blockzitat',
-        border: 'Rahmen',
-        button_apply: 'Anwenden',
-        button_cancel: 'Abbrechen',
-        button_insert: 'Einfügen',
-        button_update: 'Aktualisieren',
-        caption: 'Beschriftung',
-        class: 'Klasse',
-        classes: 'Klassen',
-        choice_no: 'Nein',
-        choice_yes: 'Ja',
-        div_create: 'Div erstellen',
-        div_edit: 'Div bearbeiten',
-        div_field_lang: 'Sprachcode',
-        div_field_title: 'Beratender Titel',
-        div_remove: 'Div entfernen',
-        div_update: 'Div-Container bearbeiten',
-        format_paragraph: 'Absatzformat',
-        format_remove: 'Formatierung entfernen',
-        height: 'Höhe',
-        history_redo: 'Wiederherstellen',
-        history_undo: 'Rückgängig machen',
-        horizontal_line_insert: 'Horizontale Linie einfügen',
-        id: 'ID',
-        indent_decrease: 'Einzug verkleinen',
-        indent_increase: 'Einzug vergrößern',
-        link: 'Link',
-        link_anchor: 'Anker',
-        link_anchor_edit: 'Anker bearbeiten',
-        link_anchor_name: 'Ankername',
-        link_anchor_properties: 'Ankereigenschaften',
-        link_anchor_remove: 'Anker entfernen',
-        link_anchor_select: 'Anker auswählen',
-        link_edit: 'Link bearbeiten',
-        link_email_address: 'E-Mail-Adresse',
-        link_email_body: 'Nachrichtentext',
-        link_email_subject: 'Betreff',
-        link_no_anchors: 'Keine Anker im Dokument verfügbar',
-        link_phone_number: 'Telefonnummer',
-        link_target: 'Zielseite',
-        link_target_new_window: 'Neues Fenster (_blank)',
-        link_target_same_window: 'Gleiches Fenster (_self)',
-        link_type: 'Link-Typ',
-        link_type_anchor: 'Link zu Anker im Text',
-        link_type_email: 'E-Mail',
-        link_type_phone: 'Telefon',
-        link_type_url: 'URL',
-        link_unlink: 'Link entfernen',
-        link_url: 'URL',
-        list_bulleted_create: 'Aufzählungsliste einfügen/entfernen',
-        list_numbered_create: 'Nummerierte Liste einfügen/entfernen',
-        modal_close: 'Schließen',
-        mode_source_code: 'Quellcode anzeigen',
-        mode_source_code_hide: 'Quellcode ausblenden',
-        padding: 'Innenabstand',
-        select: 'Auswählen',
-        spacing: 'Abstand',
-        special_characters: 'Sonderzeichen',
-        style: 'Stijl',
-        style_inline: 'Inline-Stil',
-        styles_format: 'Formatvorlagen',
-        summary: 'Zusammenfassung',
-        table_cell: 'Zelle',
-        table_cell_clear: 'Zelle(n) leeren',
-        table_cell_insert_after: 'Zelle danach einfügen',
-        table_cell_insert_before: 'Zelle davor einfügen',
-        table_cell_merge: 'Zellen verbinden',
-        table_cell_properties: 'Zelleneigenschaften',
-        table_cell_span_columns: 'Spalten zusammenfassen',
-        table_cell_span_rows: 'Zeilen zusammenfassen',
-        table_cell_split: 'Zelle teilen',
-        table_cell_type: 'Zellentyp',
-        table_cell_type_data: 'Daten',
-        table_cell_type_header: 'Überschrift',
-        table_column: 'Spalte',
-        table_column_delete: 'Spalte(n) löschen',
-        table_column_insert_after: 'Spalte danach einfügen',
-        table_column_insert_before: 'Spalte davor einfügen',
-        table_columns: 'Spalten',
-        table_delete: 'Tabelle löschen',
-        table_headers: 'Überschriften',
-        table_headers_both: 'Beide',
-        table_headers_first_column: 'Erste Spalte',
-        table_headers_first_row: 'Erste Zeile',
-        table_insert: 'Tabelle einfügen',
-        table_properties: 'Tabelleneigenschaften',
-        table_row: 'Zeile',
-        table_row_delete: 'Zeile(n) löschen',
-        table_row_insert_after: 'Zeile danach einfügen',
-        table_row_insert_before: 'Zeile davor einfügen',
-        table_rows: 'Zeile',
-        text_bold: 'Fett',
-        text_italic: 'Kursiv',
-        text_strike: 'Durchgestrichen',
-        text_subscript: 'Tiefgestellt',
-        text_superscript: 'Hochgestellt',
-        tools_maximize: 'Maximieren',
-        tools_minimize: 'Minimieren',
-        tools_show_Blocks: 'Blöcke anzeigen',
-        width: 'Breite',
-        word_wrap: 'Zeilenumbruch'
-    }
-}
+    translations[locale][key] ?? translations.en[key] ?? key
